pyproject.toml
[project]
name = "smpsave"
version = "0.0.1"
dependencies = [
    "discord.py>=2.3.2",
	"linode_api4>=5.12.1",
	"click>=8.1.7",
    'importlib-metadata; python_version<"3.10"',
]

[project.scripts]
smpsave-cli = "smpsave.cli:run_cli"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["smpsave*"]
exclude = ["smpsave.test.*"]
